creating mycontroller
hey so I've made a control class.  If you click on a node there should be a little gray box with buttons in the upper left.  I've made some other changes to node and tree, which I think I've marked.  I will add some more stuff to control so that when you mouse over a node you can see it's module composition.